Added :include and :exclude keys to :codox entry in project.clj.
:codox {:include [codox.utils codox.main]}

or

:codox {:exclude codox.main}

namespace names can be symbols or strings. You can supply a single namespace to either (as in the second example) or some collection (as in the first).

Exclusions are filtered out of all namespaces found under :sources; then, if specific inclusions have been mentioned, all namespaces that AREN'T included are filtered out as well.